requirements.txt
# langchain
# langchain-community
# langchain-google-genai
# langchain-openai
pygithub
python-dotenv
requests
fastapi
uvicorn
supabase
pydantic[email]
python-multipart
python-pptx